# apps/web/public/generated/openapi/openfoundry.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenFoundry API",
    "version": "0.1.0",
    "description": "Versioned OpenFoundry JSON/HTTP contract generated from proto services and curated REST overlays."
  },
  "servers": [
    {
      "url": "/",
      "description": "OpenFoundry API gateway root"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/datasets/create-dataset": {
      "post": {
        "summary": "DatasetService CreateDataset",
        "description": "Generated from `open_foundry.dataset` RPC `CreateDataset` in service `DatasetService`.",
        "operationId": "open_foundry.dataset.DatasetService.CreateDataset",
        "tags": [
          "open_foundry.dataset"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatasetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dataset"
                }
              }
            }
          },
          "default": {
            "description": "Structured error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-openfoundry-sdk-namespace": "dataset",
        "x-openfoundry-api-version": "v1",
        "x-openfoundry-mcp-tool": "openfoundry.dataset.createdataset",
        "x-openfoundry-stability": "beta"
      }
    },
...
    "/api/v1/datasets/get-dataset": {
      "get": {
        "summary": "DatasetService GetDataset",
        "description": "Generated from `open_foundry.dataset` RPC `GetDataset` in service `DatasetService`.",
        "operationId": "open_foundry.dataset.DatasetService.GetDataset",
        "tags": [
          "open_foundry.dataset"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "description": "Query parameter derived from `GetDatasetRequest.id`.",
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dataset"
                }
              }
            }
          },
          "default": {
            "description": "Structured error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-openfoundry-sdk-namespace": "dataset",
        "x-openfoundry-api-version": "v1",
        "x-openfoundry-mcp-tool": "openfoundry.dataset.getdataset",
        "x-openfoundry-stability": "beta"
      }
    },
...
    "/api/v1/datasets/preview-data": {
      "post": {
        "summary": "DatasetService PreviewData",
        "description": "Generated from `open_foundry.dataset` RPC `PreviewData` in service `DatasetService`.",
        "operationId": "open_foundry.dataset.DatasetService.PreviewData",
        "tags": [
          "open_foundry.dataset"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewDataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewDataResponse"
                }
              }
            }
          },
          "default": {
            "description": "Structured error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-openfoundry-sdk-namespace": "dataset",
        "x-openfoundry-api-version": "v1",
        "x-openfoundry-mcp-tool": "openfoundry.dataset.previewdata",
        "x-openfoundry-stability": "beta"
      }
    },
...
    "/api/v1/media-set/abort-transaction": {
      "post": {
        "summary": "MediaSetService AbortTransaction",
        "description": "Generated from `open_foundry.media_set` RPC `AbortTransaction` in service `MediaSetService`.",
        "operationId": "open_foundry.media_set.MediaSetService.AbortTransaction",
        "tags": [
          "open_foundry.media_set"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "default": {
            "description": "Structured error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-openfoundry-sdk-namespace": "media_set",
        "x-openfoundry-api-version": "v1",
        "x-openfoundry-mcp-tool": "openfoundry.mediaSet.aborttransaction",
        "x-openfoundry-stability": "beta"
      }
    },
    "/api/v1/media-set/commit-transaction": {
      "post": {
        "summary": "MediaSetService CommitTransaction",
        "description": "Generated from `open_foundry.media_set` RPC `CommitTransaction` in service `MediaSetService`.",
        "operationId": "open_foundry.media_set.MediaSetService.CommitTransaction",
        "tags": [
          "open_foundry.media_set"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "default": {
            "description": "Structured error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-openfoundry-sdk-namespace": "media_set",
        "x-openfoundry-api-version": "v1",
        "x-openfoundry-mcp-tool": "openfoundry.mediaSet.committransaction",
        "x-openfoundry-stability": "beta"
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbortTransactionRequest": {
        "type": "object",
        "properties": {
          "transaction_rid": {
            "type": "string"
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {}
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ArrayParams": {
        "type": "object",
        "properties": {
          "array_sub_type": {
            "$ref": "#/components/schemas/Field"
          }
        }
      },
...
      "Column": {
        "type": "object",
        "properties": {
          "data_type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "CommitTransactionRequest": {
        "type": "object",
        "properties": {
          "transaction_rid": {
            "type": "string"
          }
        }
      },
...
      "CreateDatasetRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "format": {
            "$ref": "#/components/schemas/DatasetFormat"
          },
          "name": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CsvOptions": {
        "type": "object",
        "properties": {
          "charset": {
            "type": "string"
          },
          "date_format": {
            "type": "string"
          },
          "delimiter": {
            "type": "string"
          },
          "escape": {
            "type": "string"
          },
          "header": {
            "type": "boolean"
          },
          "null_value": {
            "type": "string"
          },
          "quote": {
            "type": "string"
          },
          "timestamp_format": {
            "type": "string"
          }
        }
      },
      "CustomMetadata": {
        "type": "object",
        "properties": {
          "csv": {
            "$ref": "#/components/schemas/CsvOptions"
          }
        }
      },
      "Dataset": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "current_version": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "format": {
            "$ref": "#/components/schemas/DatasetFormat"
          },
          "id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "name": {
            "type": "string"
          },
          "owner_id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "row_count": {
            "type": "integer",
            "format": "int64"
          },
          "schema": {
            "$ref": "#/components/schemas/Schema"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "storage_path": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DecimalParams": {
        "type": "object",
        "properties": {
          "precision": {
            "type": "integer",
            "format": "int32"
          },
          "scale": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Field": {
        "type": "object",
        "properties": {
          "array": {
            "$ref": "#/components/schemas/ArrayParams"
          },
          "decimal": {
            "$ref": "#/components/schemas/DecimalParams"
          },
          "field_type": {
            "$ref": "#/components/schemas/FieldType"
          },
          "map": {
            "$ref": "#/components/schemas/MapParams"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "nullable": {
            "type": "boolean"
          },
          "struct": {
            "$ref": "#/components/schemas/StructParams"
          }
        }
      },
...
      "MapParams": {
        "type": "object",
        "properties": {
          "map_key_type": {
            "$ref": "#/components/schemas/Field"
          },
          "map_value_type": {
            "$ref": "#/components/schemas/Field"
          }
        }
      },
...
      "PreviewDataRequest": {
        "type": "object",
        "properties": {
          "dataset_id": {
            "$ref": "#/components/schemas/Uuid"
          },
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PreviewDataResponse": {
        "type": "object",
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Column"
            }
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Row"
            }
          },
          "total_rows": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Row": {
        "type": "object",
        "properties": {
          "values": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "Schema": {
        "type": "object",
        "properties": {
          "custom_metadata": {
            "$ref": "#/components/schemas/CustomMetadata"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Field"
            }
          },
          "file_format": {
            "$ref": "#/components/schemas/FileFormat"
          }
        }
      },
...
      "StructParams": {
        "type": "object",
        "properties": {
          "sub_schemas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Field"
            }
          }
        }
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "branch": {
            "type": "string"
          },
          "closed_at": {
            "type": "string",
            "format": "date-time"
          },
          "media_set_rid": {
            "type": "string"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time"
          },
          "opened_by": {
            "type": "string"
          },
          "rid": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/TransactionState"
          }
        }
      },
...
      "Uuid": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token accepted by the OpenFoundry gateway. Session and personal access tokens share this scheme."
      }
    }
  }
}